[ESRTEST-30704] Take Script out of QueryBuilder namespace
Moves the Script class out of the QueryBuilder namespace and leaves it
directly under Elasticsearch. This is done because there are scripted
elements in Elasticsearch outside the context of queries, such as
scripts that run when documents are updated or reindexed.

Therefore it makes sense to have the Script class in a more general
namespace.
