+++ lib/jay_api/elasticsearch.rb
@@ -13,6 +13,7 @@ require_relative 'elasticsearch/mixins'
 require_relative 'elasticsearch/query_builder'
 require_relative 'elasticsearch/query_results'
 require_relative 'elasticsearch/response'
+require_relative 'elasticsearch/script'
 require_relative 'elasticsearch/search_after_results'
 require_relative 'elasticsearch/stats'
 require_relative 'elasticsearch/tasks'
